// src/services/interfaces/TrendingShowResponse.ts
export interface TrendingShowResponse {
  'id': number
  'vote_count': number
  'vote_average': number
  'name': string
  'first_air_date': string
  'original_language': string
  'original_name': string
  'genre_ids': number[]
  'backdrop_path': string
  'overview': string
  'poster_path': string
  'popularity': number
  'media_type': string
}